objekt-orientierte-scriptsprachen/encadrement/pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "encadrement"
version = "0.1.0"
description = "Un petit programme de traitement d'images"
authors = ["Mara Schulke <mara.schulke@th-brandenburg.de>"]

[tool.poetry.scripts]
encadrement = "encadrement.encadrement:main"

[tool.poetry.dependencies]
python = "^3.10"
pillow = "^10.1.0"
ttkbootstrap = "^1.10.1"
